+++ src/background.rs
@@ -61,6 +61,7 @@ where
         self.inner.name()
     }
 
+    #[allow(deprecated)]
     fn threads(&self) -> Option<usize> {
         self.inner.threads()
     }